Fix USDC roundtrip for BlendShape inbetween `weight` AttrMeta
stage-converter.cc was writing `weight` to the crate as `double`
(matching AttrMetas::get_weight()'s return type), but the reader at
usdc-reader-property.cc:417 unpacks it as `float` to match the pxr
canonical encoding for usdSkel BlendShape inbetweens. Result:
USDA -> USDC -> reparse failed with "`weight` field is not `float` type"
on inbetween-blendshape-001.

Cast to float on write so the wire type matches both the pxr canonical
and our existing reader. Internal storage stays double (no API change).

Closes the long-standing usdc-roundtrip-test failure introduced by
2fbe7832 (USDC allowedTokens roundtrip).

ctest 13/13 pass, 607 unit tests pass, 354/359 USDA roundtrip
equivalent (5 unchanged Pixar-side XFAIL).

+++ src/stage-converter.cc
@@ -2168,8 +2168,11 @@ bool CrateWriter::ConvertAttributeToFields(
     attr_fields.push_back({"elementSize", v});
   }
   if (metas.has_weight()) {
+    // pxr crate stores BlendShape inbetween `weight` as float; our
+    // AttrMetas keeps it as double. Cast to float on write so the
+    // reader's float-typed unpack matches the pxr canonical.
     crate::CrateValue v;
-    v.Set(metas.get_weight());
+    v.Set(static_cast<float>(metas.get_weight()));
     attr_fields.push_back({"weight", v});
   }
   if (metas.has_comment()) {
